% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{choose_malevnc_dataset}
\alias{choose_malevnc_dataset}
\alias{choose_flyem_dataset}
\title{Choose active male VNC / FlyEM dataset}
\usage{
choose_malevnc_dataset(dataset = NULL, set = TRUE, use_clio = NA)

choose_flyem_dataset(dataset = getOption("malevnc.dataset"), set = TRUE)
}
\arguments{
\item{dataset}{Character vector specifying dataset name. Default value of
\code{NULL} will respect package options or choose default dataset if none
set.}

\item{set}{Whether to set the relevant package options or just to return a
list of the required options.}

\item{use_clio}{Whether to use clio to list datasets (expert use only;
default of \code{use_clio=NA} should do the right thing).}
}
\value{
A list of options (when \code{set=FALSE}) or the \emph{previous}
  value of those options, mimicking the behaviour of \code{\link{options}}.
}
\description{
\code{choose_malevnc_dataset} chooses the dataset that will be
  used by the male VNC package.

\code{choose_flyem_dataset} is a generic function underneath
  \code{choose_malevnc_dataset}. It is intended primarily for developers of
  other packages.
}
\details{
For \code{choose_malevnc_dataset} the \code{dataset} will be one of
  \code{MANC} (release) or \code{VNC} (pre-release) dataset. Note the latter
  will only be available for authenticated collaborators of the FlyEM team.
  When \code{dataset=NULL} the options("malevnc.dataset") will be checked,
  then the environment variable \code{MALEVNC_DATASET} and finally the
  default value (MANC i.e. released dataset) will be selected.
}
\examples{
\dontrun{
# switch dataset
op <- choose_malevnc_dataset("MANC")
on.exit(options(op))
# reset the previous state
options(op)
}
}
