% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{colour_leg_muscles}
\alias{colour_leg_muscles}
\alias{leg_muscle_palette}
\title{Create SVG figure with leg muscles coloured based on supplied colour palette}
\usage{
colour_leg_muscles(f = NULL, colpal = NULL, colids = names(colpal))

leg_muscle_palette()
}
\arguments{
\item{f}{Path to write out SVG file}

\item{colpal}{A vector of colours named by the leg muscles}

\item{colids}{The names of the muscles}
}
\value{
When \code{f} is missing, an object of class \code{xml_document} that
  can be further modified using the \code{xml2} package. Otherwise, the path
  \code{f}, invisibly.
}
\description{
\code{leg_muscle_palette} returns the default palette for the
  leg muscles.
}
\examples{
\donttest{
pal=leg_muscle_palette()
pal
pal[]='white'
pal['Fe-reductor']='red'
\dontrun{
colour_leg_muscles(f='Fe-reductor.svg', colpal=pal)
}
}
}
