% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{download_manc_registrations}
\alias{download_manc_registrations}
\title{Download MANC to JRC registrations}
\usage{
download_manc_registrations()
}
\description{
It downloads publicly available registrations:
MANC to JRC2018VNCM \url{https://figshare.com/s/7f003353c24741136df3}
JRC2018VNCM <> JRC2018VNCU \url{https://figshare.com/s/42ad71eb14e7dd51e81a}
RC2018VNCF <> JRC2018VNCU \url{https://figshare.com/s/c4589cef9180e1dd4ee1}
}
\details{
It requires installation of a suggested dependency \code{nat.jrcbrains}.
These transformations operate in units of microns.
Therefore, any point coordinates must be converted to micron units
before applying.
}
\references{
See Bogovic et al. (2018) \url{https://doi.org/10.1101/376384}
}
