% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dr_manc}
\alias{dr_manc}
\title{Status report for the malevnc package}
\usage{
dr_manc()
}
\value{
Called for its side effect
}
\description{
\code{dr_manc} gives a very useful status report for the malevnc
  package.
}
\examples{
\dontrun{
dr_manc()
}
}
