% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvidtools.R
\name{dvid_tools}
\alias{dvid_tools}
\alias{install_dvid_tools}
\title{Interface to the python dvid_tools module from Philipp Schlegel}
\usage{
dvid_tools(user = getOption("malevnc.dvid_user"), node = "neutu")

install_dvid_tools()
}
\arguments{
\item{user}{Default DVID user names}

\item{node}{default DVID node to use for queries (see
\code{\link{manc_dvid_node}})}
}
\description{
\code{dvid_tools} provides a lower level and less specific
  interface to DVID that corresponding \code{manc_*} functions but does
  include some functionality that is not yet available.

\code{install_dvid_tools} installs the python dvid_tools module
}
\details{
The dvid_tools_module will be cached using \code{memoise}. You can
  modify the default user with the `$setup` method.
}
\examples{
\dontrun{
dt=dvid_tools()
dt
# nb must explicity use integers for bodyids
dt$get_annotation(bodyid = 10000L)

# get detailed help
reticulate::py_help(dt$get_adjacency)
}
}
\seealso{
\url{https://github.com/flyconnectome/dvid_tools},
  \url{https://github.com/flyconnectome/dvid_tools},
  \url{https://emdata5.janelia.org/api/help/} for further details.
}
