% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-shortening.R
\name{flyem_shorten_url}
\alias{flyem_shorten_url}
\alias{flyem_expand_url}
\title{Shorten a Neuroglancer URL using the Janelia FlyEM link shortener}
\usage{
flyem_shorten_url(url, filename = NULL, ...)

flyem_expand_url(url, return = c("url", "json", "parsed"), ...)
}
\arguments{
\item{url}{One or more urls to shorten or expand}

\item{filename}{An optional filename to use in the shortened URL (not yet
supported)}

\item{...}{Additional arguments passed to \code{httr::POST}}

\item{return}{When expanding, whether to return a long URL, an R list or a
JSON text fragment.}
}
\value{
For \code{flyem_shorten_url} a character vector containing a short
  URL. For \code{flyem_expand_url} see a character vector or list depending
  on \code{return} argument. If the input \code{url} argument is named vector
  of length>1, then the output will also be named.
}
\description{
Shorten a Neuroglancer URL using the Janelia FlyEM link shortener
}
\details{
see
  \href{https://flyem-cns.slack.com/archives/C01BZB05M8C/p1669646269799509}{FlyEM
   CNS Slack} for more details.
}
\examples{
\dontrun{
su=flyem_shorten_url(manc_scene('group:10200'))
lu=flyem_expand_url(su)
fafbseg::ngl_decode_scene(lu)
# these give you the same result
browseURL(su)
browseURL(lu)
}
}
