% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malevnc-package.R
\docType{package}
\name{malevnc-package}
\alias{malevnc}
\alias{malevnc-package}
\title{malevnc: Support for Working with Janelia FlyEM Male VNC Dataset}
\description{
Provides support for analysis of the male Drosophila ventral nerve cord dataset released in June 2023 by the Janelia FlyEM team in collaboration with the Hess lab, Google and Drosophila Connectomics Group, Cambridge. This includes reading of connectivity information, skeleton and mesh representations of neurons and symmetrising and inter-template registrations.
}
\section{Package Options}{
 \itemize{

  \item \code{malevnc.server} The https URL to the main Janelia server
  hosting male VNC data.

  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/malevnc}
  \item Report bugs at \url{https://github.com/natverse/malevnc/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Dominik Krzeminski \email{dkk33@cam.ac.uk} (\href{https://orcid.org/0000-0003-4568-0583}{ORCID})
}

}
\keyword{internal}
