% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\docType{data}
\name{MANC.surf}
\alias{MANC.surf}
\alias{JRCFIBVNC2020MNP.surf}
\title{Surface model of neuropils in the MANC dataset}
\description{
The \code{\link{MANC.surf}} model was generated in Amira by Kazunori Shinomiya and shared with Greg 1 October 2020. It should be regarded as provisional
}
\examples{
\dontrun{
# Depends on nat
library(nat)
plot3d(MANC.surf)
}
}
