% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MANC.tracts}
\alias{MANC.tracts}
\title{Meshes of 22 MALEVNC tracts.}
\format{
An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
MANC.tracts
}
\description{
\code{MANC.tracts} contains mesh of 22 MANC tracts in
non-symmetric space:
CVL_L, CVL_R, DLT_L, DLT_R, DLV_L, DLV_R, DMT_L, DMT_R, ITD_L, ITD_R,
MDA_L, MDA_R, type-I_MTD_L, type-I_MTD_R, type-II_MTD_L, type-II_MTD_R,
type-III_MTD_L, type-III_MTD_R, VLT_L, VLT_R, VTV_L, VTV_R
}
\examples{
\dontrun{
# Depends on nat
library(nat)
plot3d(MANC.tracts)
}
}
\keyword{datasets}
