% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{manc_annotate_body}
\alias{manc_annotate_body}
\title{Set Clio body annotations}
\usage{
manc_annotate_body(
  x,
  test = TRUE,
  version = NULL,
  write_empty_fields = FALSE,
  chunksize = 50,
  ...
)
}
\arguments{
\item{x}{A data.frame, list or JSON string containing body annotations}

\item{test}{Whether to use the test clio store (recommended until you are
sure you know what you are doing).}

\item{version}{Optional clio version to associate with this annotation. The
default \code{NULL} uses the current version returned by the API.}

\item{write_empty_fields}{When \code{x} is a data.frame, this controls
whether empty fields in \code{x} overwrite fields in the clio-store
database. The (conservative) default \code{FALSE} does not overwrite. If
you do want to set to an empty value (usually the empty string) then you
must set to \code{write_empty_fields=TRUE}.}

\item{chunksize}{When you have many bodies to annotate the request will by
default be sent 50 at a time to avoid any issue with timeouts. You can
increase for a small speed up if you find your setup is fast enough. Set to
\code{Inf} to insist that all records are sent in a single request.
\bold{NB only applies when \code{x} is a data.frame}.}

\item{...}{Additional parameters passed to \code{pbapply::\link{pbsapply}}}
}
\value{
\code{NULL} invisibly on success. Errors out on failure.
}
\description{
Set Clio body annotations
}
\details{
Clio body annotations are stored in a
  \href{https://cloud.google.com/firestore}{Google Firestore} database.
  Further details are provided in
  \href{https://docs.google.com/document/d/14wzFX6cMf0JcR0ozf7wmufNoUcVtlruzUo5BdAgdM-g/edit}{basic
   docs from Bill Katz}. Each body has an associated JSON list containing a
  set of standard user visible fields. Some of these are constrained. See
  \href{https://docs.google.com/spreadsheets/d/1v8AltqyPCVNIC_m6gDNy6IDK10R6xcGkKWFxhmvCpCs/edit?usp=sharing}{Clio
   fields Google sheet} for details.

  It can take some time to apply annotations, so requests are chunked by
  default
}
\section{Validation}{
 Validation depends on how you provide your input data.
  If \code{x} is a data.frame then each row is checked for some basics including the presence of a bodyid, and empty fields are removed. In future we will also check
}

\examples{
\dontrun{
# if you give a list then you are responsible for validation
# note use of test server
manc_annotate_body(list(bodyid=10002, class='Descending Neuron',
  description='Giant Fiber'), test=TRUE)
}
}
\seealso{
Other manc-annotation: 
\code{\link{manc_body_annotations}()},
\code{\link{manc_meta}()},
\code{\link{manc_point_annotations}()}
}
\concept{manc-annotation}
