% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{manc_annotate_soma}
\alias{manc_annotate_soma}
\title{Annotate an XYZ location in Clio as a soma or root}
\usage{
manc_annotate_soma(
  pos,
  tag = c("soma", "tosoma", "root"),
  user = getOption("malevnc.clio_email"),
  description = NULL,
  ...
)
}
\arguments{
\item{pos}{XYZ positions in any raw MANC voxel coordinates in any form
compatible with \code{\link{xyzmatrix}}}

\item{tag}{One tag (or as many as there are points) to define the annotation.
See details.}

\item{user}{One user (or as many users as there are points) to define the
annotation. See details.}

\item{description}{Either \code{NULL}, one value or as any as there are
points.}

\item{...}{Additional arguments passed to \code{pbapply::\link{pbmapply}} and
then to the private \code{clio_fetch} function.}
}
\value{
the key for the new point annotation, invisibly
}
\description{
Annotate an XYZ location in Clio as a soma or root
}
\details{
You should use the tags in the following order of preference
  \itemize{

  \item{soma}{ If there is a visible soma, use this tag.}

  \item{tosoma}{ If the soma is not in the dataset, but you know where it is,
  use this tag.}

  \item{root}{ If you are not sure where the soma is, but you want to specify
  a position to act as the root e.g. skeletonising the neuron, use this tag.}

  }

  When annotating multiple points, you can provide either one user/tag or you
  must provide a user tag for every point.
}
\examples{
\dontrun{
manc_annotate_soma(cbind(31180, 17731, 31252))
}
}
