% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvidtools.R
\name{manc_check_group_complete}
\alias{manc_check_group_complete}
\title{Check if group is complete}
\usage{
manc_check_group_complete(group_id, body_ids, conn = manc_neuprint())
}
\arguments{
\item{group_id}{numeric/character with group id}

\item{body_ids}{vector with body ids to compare}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}
}
\value{
logical with \code{TRUE} if group is complete
}
\description{
In other words in all neurons from group are represented in \code{body_ids}
vector.
}
