% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_dvid_annotations}
\alias{manc_dvid_annotations}
\title{Return all DVID body annotations}
\usage{
manc_dvid_annotations(
  node = manc_dvid_node("neutu"),
  rval = c("data.frame", "list")
)
}
\arguments{
\item{node}{A DVID node as returned by \code{\link{manc_dvid_node}}. The
default is to return the current active (unlocked) node being used through
neutu.}

\item{rval}{Whether to return a fully parsed data.frame (the default) or an R
list. The data.frame is easier to work with but typically includes NAs for
many values that would be missing in the list.}
}
\value{
A \code{tibble} containing with columns including \itemize{

  \item bodyid

  \item status

  \item user

  \item naming_user

  \item instance

  \item status_user

  \item comment }

  NB these are slightly modified from
}
\description{
Return all DVID body annotations
}
\details{
See
  \href{https://flyem-cns.slack.com/archives/C01BT2XFEEN/p1619201195032400}{this
   Slack post} from Stuart Berg for details.
}
\examples{
\donttest{
mdf=manc_dvid_annotations()
head(mdf)
table(mdf$status)

\dontrun{
# compare live body annotations with version in clio
mdf.clio=manc_dvid_annotations(manc_dvid_node('clio'))
waldo::compare(mdf.clio, mdf)
}
}
}
