% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvid.R
\name{manc_dvid_node}
\alias{manc_dvid_node}
\alias{manc_dvid_nodeinfo}
\title{Information about DVID nodes / return latest node}
\usage{
manc_dvid_node(type = c("clio", "neutu", "neuprint", "master"), cached = TRUE)

manc_dvid_nodeinfo(cached = TRUE)
}
\arguments{
\item{type}{Whether to return the latest committed node (clio) or the active
node being edited in neutu (the very latest) or the node in neuprint (a
committed node that may lag behind clio). master is an alias for neutu.}

\item{cached}{Whether to return a cached value (updated every hour) or to
force a new query.}
}
\value{
A UUID string
}
\description{
\code{manc_dvid_node} returns the latest DVID node in use with a
  specific tool.

\code{manc_dvid_nodeinfo} returns a data.frame with information
  about the DVID nodes available for the male VNC dataset.
}
\examples{
\donttest{
manc_dvid_node()
manc_dvid_node('neutu')
# force
manc_dvid_node('neuprint', cached=FALSE)
}
}
