% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_ids}
\alias{manc_ids}
\title{Flexible specification of manc body ids}
\usage{
manc_ids(
  x,
  mustWork = TRUE,
  as_character = TRUE,
  integer64 = FALSE,
  conn = manc_neuprint(),
  ...
)
}
\arguments{
\item{x}{A vector of body ids, data.frame (containing a bodyid column) or a
neuroglancer URL.}

\item{mustWork}{Whether to insist that at least one valid id is returned
(default \code{TRUE})}

\item{as_character}{Whether to return segments as character rather than
numeric vector (the default is character for safety).}

\item{integer64}{whether to return ids with class bit64::integer64.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link[neuprintr]{neuprint_login}} for details.}

\item{...}{Additional arguments passed to \code{\link[neuprintr]{neuprint_search}}}
}
\description{
Flexible specification of manc body ids
}
\examples{
\donttest{
manc_ids("Giant Fiber")
}
}
\seealso{
Other manc-neuprint: 
\code{\link{manc_connection_table}()},
\code{\link{manc_neuprint}()}
}
\concept{manc-neuprint}
