% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{manc_islatest}
\alias{manc_islatest}
\title{Check if a bodyid still exists in the specified DVID node}
\usage{
manc_islatest(ids, node = manc_dvid_node("neutu"), ...)
}
\arguments{
\item{ids}{A set of body ids}

\item{node}{A DVID node (defaults to the current neutu node, see
\code{\link{manc_dvid_node}})}

\item{...}{Additional arguments passed to \code{\link{pbsapply}} and then
eventually to \code{httr::\link{HEAD}}}
}
\value{
A logical vector indicating if the body exists (TRUE) or not (FALSE).
}
\description{
Check if a bodyid still exists in the specified DVID node
}
\details{
Note that this is still quite slow (5-10 bodies per second from Cambridge)
}
\examples{
manc_islatest(10056)

\dontrun{
manc_islatest(10056:10058)

manc_islatest(10056:10058, manc_dvid_node("clio"))
}
}
