% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{manc_leg_summary}
\alias{manc_leg_summary}
\alias{manc_side_summary}
\title{Simple summaries of which regions different neurons innervate}
\usage{
manc_leg_summary(ids, long = FALSE, conn = manc_neuprint())

manc_side_summary(ids, long = FALSE, conn = manc_neuprint())
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{long}{Whether to return results in wide (default) or long format.}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}
}
\value{
a data.frame with one row per neuron (when \code{long=FALSE}) or one
  row per roi/IO combination (when \code{long=TRUE}).
}
\description{
\code{manc_leg_summary} summarises I/O in the main leg
  neuropils.

\code{manc_side_summary} summarises connections within all of
  the ROIs that have an L or R designation.
}
\examples{
dnals=manc_leg_summary(c(10126, 10118))
dnals
manc_leg_summary(c(10126, 10118), long=T)
heatmap(data.matrix(dnals[grep("_down", colnames(dnals))]),
  Colv = NA, scale = 'none')

\donttest{
dnls=manc_leg_summary('class:Descending')
heatmap(data.matrix(dnls[grep("_down", colnames(dnls))]),
  Colv = NA, scale = 'none')
}
manc_side_summary('Giant Fiber')
}
