% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{manc_lr_position}
\alias{manc_lr_position}
\title{Calculate the left-right position wrt to the symmetrised MANC midline}
\usage{
manc_lr_position(x, ...)
}
\arguments{
\item{x}{An object containing XYZ vertex locations in microns, compatible
with \code{\link{xyzmatrix}}}

\item{...}{Additional arguments passed to \code{\link{xform}}}
}
\value{
A vector of point displacements in microns where 0 is at the midline
  and positive values are to the fly's right.
}
\description{
Calculate the left-right position wrt to the symmetrised MANC midline
}
\examples{
library(nat)
lr=manc_lr_position(xyzmatrix(mancsomapos)/125)
# red for left, green for right (nautical convention)
points3d(xyzmatrix(mancsomapos), col=ifelse(lr<0, "red","green"))
plot3d(boundingbox(mancsomapos))
}
