% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_meta}
\alias{manc_meta}
\title{Return full metadata from Clio/DVID for MANC bodyids (cached by default)}
\usage{
manc_meta(
  ids = NULL,
  cache = TRUE,
  unique = FALSE,
  node = "neutu",
  update.bodyids = FALSE
)
}
\arguments{
\item{ids}{bodyids specified in any form accepted by \code{\link{manc_ids}}}

\item{cache}{Whether to use a cache with a 5 min lifetime, default
\code{TRUE}.}

\item{unique}{Whether to remove any duplicate ids from the results.}

\item{node}{A DVID node as returned by \code{\link{manc_dvid_node}}. The
default is to return the current active (unlocked) node being used through
neutu.}

\item{update.bodyids}{Whether to update the bodyid associated with
annotations based on the position field. The default value of this has been
switched to \code{FALSE} as of Feb 2022.}
}
\value{
A data.frame with columns determined by
  \code{\link{manc_dvid_annotations}} and
  \code{\link{manc_body_annotations}}. All columns from
  \code{manc_dvid_annotations} will have the prefix \code{dvid_} (except for
  the single \code{bodyid} column). In addition a column \code{dvid_group}
  will be generated by parsing the \code{dvid_instance} column. This will
  have more up to date LR match information than the clio \code{group} field.
}
\description{
Return full metadata from Clio/DVID for MANC bodyids (cached by default)
}
\details{
This function will the latest per bodyid metadata using
  \code{\link{manc_dvid_annotations}} and
  \code{\link{manc_body_annotations}}. Since obtaining this information can
  take 5-10s even for a single id, it is by default cached with a 5 min
  timeout. The second call for the same request should be more or less
  instantaneous, while a different set of ids should still return much
  faster. Annotations are returned for the 'neutu' i.e. latest DVID node by
  default.
}
\examples{
\dontrun{
manc_meta('DNp01')
manc_meta(10002)
}
}
\seealso{
Other manc-annotation: 
\code{\link{manc_annotate_body}()},
\code{\link{manc_body_annotations}()},
\code{\link{manc_point_annotations}()}
}
\concept{manc-annotation}
