% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_neuprint}
\alias{manc_neuprint}
\title{Login to MANC neuprint server}
\usage{
manc_neuprint(
  server = getOption("malevnc.neuprint"),
  dataset = getOption("malevnc.neuprint_dataset"),
  token = Sys.getenv("neuprint_token"),
  ...
)
}
\arguments{
\item{server}{A neuprint server (a sensible default should be chosen)}

\item{dataset}{A neuprint dataset (a sensible default should be chosen based
on which dataset has been specified for Clio/DVID)}

\item{token}{neuprint authorisation token obtained e.g. from
neuprint.janelia.org website.}

\item{...}{Additional arguments passed to \code{\link{neuprint_login}}}
}
\value{
a \code{\link{neuprint_connection}} object returned by
  \code{\link{neuprint_login}}
}
\description{
Login to MANC neuprint server
}
\examples{
\donttest{
vncc=manc_neuprint()
anchorids <- neuprintr::neuprint_ids("status:Anchor", conn=vncc)
# the last connection will be used by default
anchormeta <- neuprintr::neuprint_get_meta("status:Anchor")

plot(cumsum(sort(anchormeta$pre, decreasing = TRUE)), ylab='Cumulative presynapses')
plot(cumsum(sort(anchormeta$post, decreasing = TRUE)), ylab='Cumulative postsynapses')
}
}
\seealso{
Other manc-neuprint: 
\code{\link{manc_connection_table}()},
\code{\link{manc_ids}()}
}
\concept{manc-neuprint}
