% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_neuprint_meta}
\alias{manc_neuprint_meta}
\title{Fetch neuprint metadata for MANC neurons}
\usage{
manc_neuprint_meta(
  ids = NULL,
  conn = manc_neuprint(),
  roiInfo = FALSE,
  fields.regex.exclude = NULL,
  fields.regex.include = NULL,
  ...
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}

\item{roiInfo}{whether to include the \code{roiInfo} field detailing synapse
numbers in different locations. This is omitted by default as it is
returned as a character vector of unprocessed JSON.}

\item{fields.regex.exclude, fields.regex.include}{Optional regular expressions
to define fields to include or exclude from the returned metadata.}

\item{...}{Additional arguments passed to \code{\link{neuprint_get_meta}}}
}
\value{
A data.frame with one row for each (unique) id and NAs for all
  columns except bodyid when neuprint holds no metadata.
}
\description{
Fetch neuprint metadata for MANC neurons
}
\details{
When \code{ids = NULL} then a default set of bodies is selected
  using the \code{\link{manc_dvid_annotations}} function. Since April 2025
  this uses the \code{node='neuprint'}. This should correspond to all neurons
  with an annotation. You can also use other searches e.g. to fetch all
  neurons, see examples.
}
\examples{
\donttest{
manc_neuprint_meta("DNp01")

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
}
\dontrun{
# fetch all neurons
allneurons <- manc_neuprint_meta('where:exists(n.bodyId)')
# in theory you could do this, but it often seems to time out:
allsegs=neuprintr::neuprint_ids('where:exists(n.bodyId)', all_segments=TRUE)
}
}
\seealso{
\code{\link{manc_ids}}
}
