% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_point_annotations}
\alias{manc_point_annotations}
\title{Return point annotations from Clio store}
\usage{
manc_point_annotations(groups = "UK Drosophila Connectomics")
}
\arguments{
\item{groups}{Defines a group for which we would like to see all annotations. When NULL, only returns annotations for your own user.}
}
\value{
A data.frame of annotations
}
\description{
Return point annotations from Clio store
}
\examples{
\dontrun{
mpa=manc_point_annotations()
head(mpa)
}
}
\seealso{
Other manc-annotation: 
\code{\link{manc_body_annotations}()}
}
\concept{manc-annotation}
