% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{manc_remove_annotations}
\alias{manc_remove_annotations}
\title{Removes records associated with given bodyids from Clio}
\usage{
manc_remove_annotations(x, test = T)
}
\arguments{
\item{x}{data.frame or vector with bodyids}

\item{test}{Whether to use the test clio store (recommended until you are
sure you know what you are doing).}
}
\description{
Use with EXTREME caution!
Only certain users have permissions to delete annotations.
This operation is irrevertible.
}
\examples{
\dontrun{
manc_remove_annotations(c(111, 222))
}
}
