% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{manc_scene}
\alias{manc_scene}
\title{Return a sample Neuroglancer scene URL for MANC dataset}
\usage{
manc_scene(server = c("appspot", "janelia"))
}
\arguments{
\item{server}{Whether to use the Google server (newest version of
neuroglancer) or Janelia server (apparently required for annotation).}
}
\value{
A character vector containing a single Neuroglancer URL.
}
\description{
Return a sample Neuroglancer scene URL for MANC dataset
}
\details{
You will need to authenticate at \url{https://neuprint.janelia.org/}
  in order to use the annotation features. I recommend logging in and out of
  neuprint if you still get authentication errors from Neuroglancer when
  attempting to use the annotation layer.

  Awkwardly the mesh and label field ROIs currently default to different
  colours.
}
\examples{
\donttest{
browseURL(manc_scene())
}
}
