% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvidtools.R
\name{manc_set_dvid_instance}
\alias{manc_set_dvid_instance}
\alias{manc_set_dvid_annotations}
\title{Set DVID annotations for one or more ids.}
\usage{
manc_set_dvid_instance(
  bodyid,
  instance = NULL,
  type = NULL,
  synonyms = NULL,
  user = getOption("malevnc.dvid_user"),
  node = "neutu",
  ...
)

manc_set_dvid_annotations(bodyid, annlist, node = "neutu")
}
\arguments{
\item{bodyid}{One or more body ids (only one for
\code{manc_set_dvid_annotations})}

\item{instance}{An instance, either a candidate type or an authoritative type with suffix to indicate side.}

\item{type}{The authoritative type for a neuron}

\item{synonyms}{An alternative type for a neuron (e.g. one used in the past literature, one which is useful but unlikely to be the authoritative type etc).}

\item{user}{The DVID user. Defaults to \code{options("malevnc.dvid_user")}.}

\item{node}{The dvid node to use (defaults to 'neutu' i.e. current open node)}

\item{...}{Additional arguments passed to \code{pbapply::pbmapply} when
passing more than one bodyid.}

\item{annlist}{An R list containing key value pairs. Key names are not
checked.}
}
\value{
logical indicating success

logical indicating success
}
\description{
\code{manc_set_dvid_instance} sets DVID type and/or instance
  (name) information for one or more ids. It also adds user information for
  each field.

\code{manc_set_dvid_annotations} is a low level function to set
  arbitrary DVID annotations for exactly one body (expert use only!)
}
\examples{
\dontrun{
manc_set_dvid_instance(1234567, type='LHAD1g1')
}
\dontrun{
manc_set_dvid_annotations(1234567, list(instance='LHAD1g1', instance_user=''))

}
}
