% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvidtools.R
\name{manc_set_lrgroup}
\alias{manc_set_lrgroup}
\title{Set LR matching groups for neurons in DVID and optionally Clio}
\usage{
manc_set_lrgroup(
  ids,
  dryrun = TRUE,
  Force = FALSE,
  Partial = FALSE,
  group = NA,
  clio = TRUE
)
}
\arguments{
\item{ids}{A set of ids belonging to the same group}

\item{dryrun}{When \code{TRUE}, the default, show what will
happen rather than applying the annotations.}

\item{Force}{Whether to update DVID instances (and clio group) even when
there is existing DVID instance information.}

\item{Partial}{Assigns only partial annotations to DVID instances
that do not yet have annotation.}

\item{group}{Set a specific group id.}

\item{clio}{Whether push to Clio or not.}
}
\description{
Set LR matching groups for neurons in DVID and optionally Clio
}
\examples{
\dontrun{
# check it makes sense, dryrun=T
manc_set_lrgroup(c(12516, 12706))
# apply
manc_set_lrgroup(c(12516, 12706), dryrun=F)
}
}
