% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{manc_size}
\alias{manc_size}
\title{Return the size (in voxels) of specified bodies}
\usage{
manc_size(ids, node = "neutu", chunksize = 5000L, ...)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{node}{A DVID node (defaults to the current neutu node, see
\code{\link{manc_dvid_node}})}

\item{chunksize}{Split requests into chunks of maximum this size}

\item{...}{Additional arguments passed to \code{\link{pbsapply}}}
}
\value{
Numeric vector of voxel sizes
}
\description{
Return the size (in voxels) of specified bodies
}
\examples{
manc_size(10056)
# zero as doesn't exist
manc_size(10000056)
\dontrun{
# try splitting up
ids=manc_size("class:Ascending Interneuron", chunksize=500L, cl=4)
}
}
\seealso{
\code{\link{manc_neuprint_meta}} which returns the size of objects
  recorded in neuprint. This is much faster when up to date.
}
