% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{manc_somapos}
\alias{manc_somapos}
\alias{manc_rootpos}
\title{Return the soma or root position of MANC bodyids}
\usage{
manc_somapos(
  ids = NULL,
  details = FALSE,
  duplicates = !details,
  somatags = c("soma", "tosoma", "root", "all"),
  node = "neutu",
  cache = TRUE,
  clio = TRUE
)

manc_rootpos(
  ids = NULL,
  details = FALSE,
  duplicates = !details,
  node = "neutu",
  cache = TRUE,
  clio = TRUE
)
}
\arguments{
\item{ids}{MANC bodyids in any form compatible with \code{\link{manc_ids}}.
The default \code{ids=NULL} returns the position of all known soma points.}

\item{details}{Whether to include details such as bodyids in the return.}

\item{duplicates}{Whether to include cases where the same bodyid has multiple
soma annotations. Default is to include duplicates when
\code{details=TRUE}.}

\item{somatags}{Which annotations to include. The default is to include both
\code{"soma"} and  \code{"tosoma"} tags (see details). \code{"all"} is an
alias for \code{c("soma", "tosoma", "root")}.}

\item{node}{The DVID node to use for XYZ to bodyid lookups. See
\code{\link{manc_xyz2bodyid}}.}

\item{cache}{whether to add a 5 min cache for the position lookups.}

\item{clio}{Whether to include point annotations from clio as well as the
contents of \code{\link{mancsomapos}}.}
}
\value{
When \code{details=FALSE} an Nx3 matrix. When \code{TRUE}, a
  data.frame containing XYZ position, bodyid, source and tag
}
\description{
\code{manc_rootpos} returns the root point of all neurons and is
  a convenience wrapper around \code{manc_somapos}.
}
\details{
Currently there are two sources of soma information, the
  \code{\link{mancsomapos}} data.frame distributed with the malevnc package
  and point annotations stored in Clio. The data.frame was the result of
  initial annotation work principally by Chris Ordish at FlyEM who marked two
  points to define a sphere. It contains some cases where the midpoint
  between the two marked points does not land on the intended soma; there are
  also cases where the segmentation inside the soma is not contiguous e.g.
  because the nucleus is a separate body.

  Clio point annotations were added after predicting a set of objects without
  an annotated soma that might be intrinsic VNC neurons based on features
  such as pre- and post-synapse numbers, size, skeleton path length etc.

  with a tag, described below.

  There have consistently been a few cases of bodies with multiple annotated
  somata. These are sometimes due to annotation error.

  \itemize{

  \item soma tags are placed only on somata visible within the dataset

  \item tosoma tags are placed on cell body fibres leading to the soma of an
  intrinsic neuron that may be missing (e.g. because of damage to the
  specimen or because of difficulty following the cell body fibre all the way
  to the soma due to segmentation issues.)

  \item root tags are placed on a part of the neuron that is likely to be the
  closest part of the neuron to the soma within the image volume. Two example
  use cases are DNs or sensory neurons whose soma is outside the volume. }
}
\examples{
\donttest{
# just the XYZ position
manc_somapos(10000)
# with details including data source
manc_somapos(10000, details=TRUE)
}
}
\seealso{
supersedes \code{\link{mancsomapos}}
}
