% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_user_annotations}
\alias{manc_user_annotations}
\title{Read point annotations from DVID using neuprint authentication}
\usage{
manc_user_annotations(
  email = "jefferis@gmail.com",
  node = manc_dvid_node(),
  raw = FALSE,
  simplifyVector = TRUE,
  neuprint_connection = NULL,
  tz = "UTC"
)
}
\arguments{
\item{email}{The google email address used to authenticate with neuprint}

\item{node}{The DVID node identifier}

\item{raw}{Whether to return the raw \code{httr::\link[httr]{GET}} response
(default \code{FALSE}) so that you can process it yourself rather than a
pre-processed R list.}

\item{simplifyVector}{Whether to simplify lists to vectors (and data frames
where appropriate). Default \code{TRUE}, see
\code{jsonlite::\link[jsonlite]{fromJSON}} for details.}

\item{neuprint_connection}{A \code{\link[neuprintr]{neuprintr}} connection
object returned by \code{\link[neuprintr]{neuprint_login}}. This includes
the required authorisation information to connect to DVID.}

\item{tz}{Time zone for edit timestamps. Defaults to "UTC" i.e. Universal
Time, Coordinated. Set to "" for your current timezone. See
\code{\link{as.POSIXct}} for more details.}
}
\value{
A data.frame
}
\description{
Read point annotations from DVID using neuprint authentication
}
\examples{
\donttest{
df=manc_user_annotations()
head(df)
json=httr::content(manc_user_annotations(raw=TRUE))
}
}
