% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.R
\name{manc_view3d}
\alias{manc_view3d}
\title{Set a standard viewpoint for MANC data}
\usage{
manc_view3d(
  viewpoint = c("ventral", "dorsal", "left", "right"),
  FOV = 0,
  template = c("MANCsym"),
  extramat = NULL,
  returnparams = FALSE,
  ...
)
}
\arguments{
\item{viewpoint}{A string specifying an anatomical viewpoint (defaults to
ventral)}

\item{FOV}{The Field of View (defaults to 0 => orthographic projection) (see
\code{\link[rgl]{par3d}} for details).}

\item{template}{The template object implied by the \code{viewpoint}
specification. Currently only the symmetric MANC template is supported.}

\item{extramat}{An optional extra transformation matrix to be applied after
the one implied by the viewpoint argument.}

\item{returnparams}{When \code{FALSE} uses \code{par3d} to change the rgl
device. When \code{TRUE} just returns the settings for you to use. See
examples.}

\item{...}{additional arguments passed to \code{\link[rgl]{par3d}}}
}
\value{
When \code{returnparams} a named list
}
\description{
Set a standard viewpoint for MANC data
}
\details{
See
  \href{https://flyem-cns.slack.com/archives/C02GY69SY3H/p1660679361976619}{slack
  discussion} for details.
}
\examples{
# default parameters
manc_view3d(returnparams=TRUE)
\dontrun{
plot3d(MANC.tissue.surf.sym)
manc_view3d("ventral")
manc_view3d("dorsal")
manc_view3d("left")

## open a new window
# with regular rgl function
open3d(manc_view3d('ventral', returnparams=TRUE))
# with nat, allowing interactive pan
nopen3d(manc_view3d('ventral', returnparams=TRUE))
plot3d(MANC.tissue.surf.sym)
}
}
