% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzpos.R
\name{manc_xyz2bodyid}
\alias{manc_xyz2bodyid}
\title{Find the bodyid for an XYZ location}
\usage{
manc_xyz2bodyid(xyz, node = manc_dvid_node("neutu"), viafile = NA)
}
\arguments{
\item{xyz}{location in raw pixels}

\item{node}{a DVID node}

\item{viafile}{whether to use a file to pass data to curl (expert use only)}
}
\value{
A character vector of body ids
}
\description{
Find the bodyid for an XYZ location
}
\examples{
ids=manc_xyz2bodyid(mancneckseeds)
tids=table(ids)
# are there many duplicate ids
table(tids)

dups=names(tids[tids>1])
\dontrun{
manc_scene(ids=dups)
}
}
