% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mancneckseeds}
\alias{mancneckseeds}
\title{Seed plane for the Neck Connective}
\format{
An object of class \code{data.frame} with 3781 rows and 3 columns.
}
\usage{
mancneckseeds
}
\description{
These were generated by Chris Ordish of FlyEM and
\href{https://flyem-cns.slack.com/archives/C01BT2XFEEN/p1618779811015700}{passed
to Greg Jefferis on 18th April 2021 by Stuart Berg}.
}
\details{
Note that these positions are raw (8nm voxel) locations.
}
\examples{
head(mancneckseeds)
}
\seealso{
\code{\link{mancsomapos}}
}
\keyword{datasets}
