% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mancsomapos}
\alias{mancsomapos}
\title{Soma locations for all intrinsic neurons of the ventral nerve cord}
\format{
An object of class \code{data.frame} with 15613 rows and 8 columns.
}
\usage{
mancsomapos
}
\description{
These were generated as neuroglancer annotations by Chris Ordish of the FlyEM
team and
\href{https://flyem-cns.slack.com/archives/C01BT2XFEEN/p1614781577010700}{passed
to Greg Jefferis on 3rd March 2021 by Pat Rivlin}. These have been superseded
by more dynamic data available from DVID and Clio. Therefore until the
situation stabilises, it is recommended to use the \code{\link{manc_somapos}}
function.
}
\details{
Note that these positions are raw (8nm voxel) locations.
}
\examples{
head(mancsomapos)

library(nat)
points3d(xyzmatrix(mancsomapos))
plot3d(boundingbox(xyzmatrix(mancsomapos)))

# nm
summary(xyzmatrix(mancsomapos)*8)
# microns
summary(xyzmatrix(mancsomapos)/125)

\dontrun{
# find bodyid for all somata
mancsomapos$bodyid=manc_xyz2bodyid(mancsomapos)
}
}
\seealso{
\code{\link{manc_somapos}}
}
\keyword{datasets}
