% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\docType{data}
\name{MANCsym}
\alias{MANCsym}
\title{MANC symmetric template}
\format{
An object of class \code{templatebrain} of dimension 672 x 830 x 1280.
}
\usage{
MANCsym
}
\description{
MANC symmetric template
}
\keyword{datasets}
