% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{parse_query}
\alias{parse_query}
\title{Returns default query list or extends the existing list}
\usage{
parse_query(query, version)
}
\description{
Returns default query list or extends the existing list
}
