% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{read_draco_meshes}
\alias{read_draco_meshes}
\alias{read_manc_meshes}
\title{Read draco encoded 3D meshes from tarballs (as used by Janelia FlyEM)}
\usage{
read_draco_meshes(x)

read_manc_meshes(id, node = "aefff")
}
\arguments{
\item{x}{A URL to a remote location or path to a file on disk}

\item{id}{A manc bodyid}

\item{node}{the dvid node containing the dataset}
}
\value{
A \code{\link{neuronlist}} object containing
  \code{rgl::\link[rgl]{mesh3d}} objects
}
\description{
the low-level \code{read_draco_meshes} reads meshes from tarballs

\code{read_manc_meshes} reads the meshes for a single MANC body id.
}
\examples{
\dontrun{
# NB you need VPN access for this
m=read_manc_meshes(23547362285)
plot3d(m)
clear3d()
plot3d(m, type='wire')
}
}
