% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malevnc-package.R
\docType{package}
\name{malevnc-package}
\alias{malevnc}
\alias{malevnc-package}
\title{malevnc: Support for working with Janelia FlyEM Male VNC dataset}
\description{
Will provide regional surface meshes and template
    brain support for the Janelia FlyEM / Hess lab male VNC data set
    currently for which proofreading started Autumn 2020.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/malevnc}
  \item Report bugs at \url{https://github.com/flyconnectome/malevnc/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
