% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MANC.tissue.surf}
\alias{MANC.tissue.surf}
\title{Simplified tissue surface of MALEVNC}
\format{
An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
MANC.tissue.surf
}
\description{
Simplified tissue surface of MALEVNC
}
\examples{
\dontrun{
# Depends on nat
library(nat)
wire3d(MANC.tissue.surf)
}
}
\keyword{datasets}
