% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_body_annotations}
\alias{manc_body_annotations}
\title{Return clio-store body annotations for set of ids or a flexible query}
\usage{
manc_body_annotations(
  ids = NULL,
  query = NULL,
  json = FALSE,
  config = NULL,
  ...
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{query}{A json query string (see examples or documentation) or an R list
with field names as elements.}

\item{json}{Whether to return unparsed JSON rather than an R list (default
\code{FALSE}).}

\item{config}{An optional httr::config (expert use only, must include a
bearer token)}

\item{...}{Additional arguments passed to \code{pbapply::\link{pblapply}}}
}
\value{
An R data.frame or a character vector containing JSON (when
  \code{json=TRUE}).
}
\description{
Return clio-store body annotations for set of ids or a flexible query
}
\details{
Missing values in each output column are filled with NA. But if a
  whole column is missing from the results of a particular query then it will
  not appear at all.

  When neither \code{query} and \code{ids} are missing then we return all
  entries in the clio store database. This currently includes annotations for
  all body ids - even the ones that are no longer current.
}
\examples{
\dontrun{
manc_body_annotations(ids=11442)
manc_body_annotations(query='{"hemilineage": "0B"}')
manc_body_annotations(query=list(user="janedoe@gmail.com"))
manc_body_annotations(query=list(soma_side="RHS"))
manc_body_annotations(ids=manc_xyz2bodyid(mancneckseeds))
# use clio node to ensure for bodyid consistency
manc_body_annotations(ids=
  manc_xyz2bodyid(mancneckseeds, node="clio"))

# fetch all bodyids
mba=manc_body_annotations()
}
}
\seealso{
\href{https://docs.google.com/document/d/14wzFX6cMf0JcR0ozf7wmufNoUcVtlruzUo5BdAgdM-g/edit}{basic
docs from Bill Katz}.

Other manc-annotation: 
\code{\link{manc_point_annotations}()}
}
\concept{manc-annotation}
