% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_body_annotations}
\alias{manc_body_annotations}
\title{Return clio-store body annotations for set of ids or a flexible query}
\usage{
manc_body_annotations(
  ids = NULL,
  query = NULL,
  json = FALSE,
  config = NULL,
  cache = FALSE,
  update.bodyids = FALSE,
  test = FALSE,
  show.extra = c("none", "user", "time", "all"),
  ...
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{query}{A json query string (see examples or documentation) or an R list
with field names as elements.}

\item{json}{Whether to return unparsed JSON rather than an R list (default
\code{FALSE}).}

\item{config}{An optional httr::config (expert use only, must include a
bearer token)}

\item{cache}{Whether to cache the result of this call for 5 minutes.}

\item{update.bodyids}{Whether to update the bodyid associated with
annotations based on the position field. The default value of this has been
switched to \code{FALSE} as of Feb 2022.}

\item{test}{Whether to unset the clio-store test server (default
\code{FALSE})}

\item{show.extra}{Extra columns to show with user/timestamp information.}

\item{...}{Additional arguments passed to \code{pbapply::\link{pblapply}}}
}
\value{
An R data.frame or a character vector containing JSON (when
  \code{json=TRUE}). Two additional fields will be added \itemize{

  \item original.bodyid When \code{update.bodyids=TRUE} this field contains
  the original bodyid from Clio whereas \code{bodyid} contains the updated
  value implied by the position.

  \item \code{auto} \code{TRUE} signals that the record contains only data
  automatically copied over from DVID without any manual annotation.

  See
  \href{https://flyem-cns.slack.com/archives/C01MYQ1AQ5D/p1628214375055400}{slack}
   for details of the position / position type fields. }
}
\description{
Return clio-store body annotations for set of ids or a flexible query
}
\details{
Missing values in each output column are filled with NA. But if a
  whole column is missing from the results of a particular query then it will
  not appear at all.

  When neither \code{query} and \code{ids} are missing then we return all
  entries in the clio store database. This currently includes annotations for
  all body ids - even the ones that are no longer current.
}
\examples{
\dontrun{
manc_body_annotations(ids=11442)
manc_body_annotations(ids=11442, show.extra='user')
manc_body_annotations(query='{"hemilineage": "0B"}')
manc_body_annotations(query=list(user="janedoe@gmail.com"))
manc_body_annotations(query=list(soma_side="RHS"))
manc_body_annotations(ids=manc_xyz2bodyid(mancneckseeds))
# use clio node to ensure for bodyid consistency
manc_body_annotations(ids=
  manc_xyz2bodyid(mancneckseeds, node="clio"))

# fetch all bodyids
mba=manc_body_annotations()
}
}
\seealso{
\href{https://clio-store-vwzoicitea-uk.a.run.app/docs}{swagger docs}
  or
\href{https://docs.google.com/document/d/14wzFX6cMf0JcR0ozf7wmufNoUcVtlruzUo5BdAgdM-g/edit}{basic
docs from Bill Katz}.

Other manc-annotation: 
\code{\link{manc_annotate_body}()},
\code{\link{manc_meta}()},
\code{\link{manc_point_annotations}()}
}
\concept{manc-annotation}
