% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_body_annotations}
\alias{manc_body_annotations}
\title{Return clio-store body annotations for set of ids or a flexible query}
\usage{
manc_body_annotations(ids = NULL, query = NULL, json = FALSE, config = NULL)
}
\arguments{
\item{ids}{A vector of one or more ids}

\item{query}{A json query string (see examples or documentation) or an R list
with field names as elements.}

\item{json}{Whether to return unparsed JSON rather than an R list (default
\code{FALSE}).}

\item{config}{An optional httr::config (expert use only, must include a
bearer token)}
}
\value{
An R data.frame or a character vector containing JSON (when
  \code{json=TRUE}).
}
\description{
Return clio-store body annotations for set of ids or a flexible query
}
\details{
Missing values in each output column are filled with NA. But if a
  whole column is missing from the results of a particular query then it will
  not appear at all.
}
\examples{
\dontrun{
manc_body_annotations(ids=11442)
manc_body_annotations(query='{"hemilineage": "0B"}')

}
}
\seealso{
\href{https://docs.google.com/document/d/14wzFX6cMf0JcR0ozf7wmufNoUcVtlruzUo5BdAgdM-g/edit}{basic
docs from Bill Katz}.
}
\concept{manc-annotation}
