% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_connection_table}
\alias{manc_connection_table}
\title{Convenience wrapper for neuprint connection queries}
\usage{
manc_connection_table(
  x,
  partners = c("inputs", "outputs"),
  prepost = c("PRE", "POST"),
  conn = manc_neuprint(),
  ...
)
}
\arguments{
\item{x}{Bodyids in any form understandable to \code{\link{manc_ids}}}

\item{partners}{Either inputs or outputs. Redundant with \code{prepost}, but
probably clearer.}

\item{prepost}{\code{PRE}: look for partners presynaptic (i.e upstream
inputs) or \code{POST}: postsynaptic (downstream outputs) to the given
\code{bodyids}}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}

\item{...}{additional arguments passed to
\code{\link{neuprint_connection_table}}}
}
\value{
see \code{\link{neuprint_connection_table}}
}
\description{
Convenience wrapper for neuprint connection queries
}
\examples{
\donttest{
down=manc_connection_table("Giant Fiber", partners='outputs')
\dontrun{
manc_scene(down$partner[1:8], open=TRUE)
}
}
}
\seealso{
Other manc-neuprint: 
\code{\link{manc_ids}()},
\code{\link{manc_neuprint}()}
}
\concept{manc-neuprint}
