% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvid.R
\name{manc_dvid_node}
\alias{manc_dvid_node}
\title{Return the latest DVID node}
\usage{
manc_dvid_node(type = c("clio", "neutu", "neuprint"), cached = TRUE)
}
\arguments{
\item{type}{Whether to return the latest committed node (clio) or the active
node being edited in neutu (the very latest) or the node in neuprint (a
committed node that may lag behind clio)}

\item{cached}{Whether to return a cached value (updated every hour) or to
force a new query.}
}
\value{
A UUID string
}
\description{
Return the latest DVID node
}
\examples{
\donttest{
manc_dvid_node()
manc_dvid_node('neutu')
# force
manc_dvid_node('neuprint', cached=FALSE)
}
}
