% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_ids}
\alias{manc_ids}
\title{Flexible specification of manc body ids}
\usage{
manc_ids(
  x,
  mustWork = TRUE,
  as_character = TRUE,
  integer64 = FALSE,
  unique = TRUE,
  cache = NA,
  conn = manc_neuprint(),
  ...
)
}
\arguments{
\item{x}{A vector of body ids, data.frame (containing a bodyid column) or a
neuroglancer URL.}

\item{mustWork}{Whether to insist that at least one valid id is returned
(default \code{TRUE})}

\item{as_character}{Whether to return segments as character rather than
numeric vector (the default is character for safety).}

\item{integer64}{whether to return ids with class bit64::integer64.}

\item{unique}{Whether to ensure that only unique ids are returned (default
\code{TRUE})}

\item{cache}{whether to cache the query. When \code{cache=NA} (the default)
queries are cached for neuprint snapshot versions (but not production
datasets). See details.}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}

\item{...}{Additional arguments passed to \code{\link{neuprint_get_meta}}}
}
\description{
\code{manc_ids} provides a convenient way to extract body ids
  from a variety of objects as well as allowing text searches against
  type/instance information defined in neuprint.
}
\details{
This function will now cache neuprint queries when using a snapshot release
  (which is assumed not to change). Snapshot releases are identified by
  containing the string \code{":v"} as in \code{manc:v1.2.3}. The cache
  currently lasts for 24h.
}
\examples{
\donttest{
# search by type
manc_ids("DNp01")
# You can also do more complex queries using regular expressions
# introduced by a slash and specifying the field to be searched
dns=manc_ids("/type:DN.+")

# you can also use Neo4J cypher queries by using the where: prefix
# note that each field of the neuron must prefixed with "n."
bignogroupids <-
  manc_ids("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
\dontrun{
# Finally you can use the same queries wherever you specify body ids
# NB if you want to be sure that regular neuprintr functions target
# the VNC dataset, use conn=manc_neuprint()
lrpairs.meta=neuprintr::neuprint_get_meta("/name:[0-9]{5,}_[LR]", conn=manc_neuprint())
}
}
}
\seealso{
\code{\link{neuprint_ids}}

Other manc-neuprint: 
\code{\link{manc_connection_table}()},
\code{\link{manc_neuprint}()}
}
\concept{manc-neuprint}
