% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{manc_islatest}
\alias{manc_islatest}
\title{Check if a bodyid still exists in the specified DVID node}
\usage{
manc_islatest(
  ids,
  node = "neutu",
  method = c("auto", "size", "sparsevol"),
  ...
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{node}{A DVID node (defaults to the current neutu node, see
\code{\link{manc_dvid_node}})}

\item{method}{Which DVID endpoint to use. Expert use only.}

\item{...}{Additional arguments passed to \code{\link{pbsapply}} and then
eventually to \code{httr::\link{HEAD}}}
}
\value{
A logical vector indicating if the body exists (TRUE) or not (FALSE).
}
\description{
Check if a bodyid still exists in the specified DVID node
}
\details{
Note that the raw speed of this call is still quite slow (typically
  5-30 bodies per second from Cambridge). When more than one body id is
  provided this actually uses \code{\link{manc_size}} to check if the body
  has >0 voxels. This is a somewhat expensive operation on the server but can
  be accept multiple ids. As a further speed-up, when many ids are requested
  and \code{method="auto"} (the recommended default) then the DVID status of
  bodies is checked. In theory only valid bodies should have such a status
  (\href{https://flyem-cns.slack.com/archives/C01MYQ1AQ5D/p1621448634011800?thread_ts=1621434072.008700&cid=C01MYQ1AQ5D}{see
  Slack message}).
}
\examples{
manc_islatest(10056)

\dontrun{
manc_islatest(10056:10058)

manc_islatest(10056:10058, "clio")
}
}
