% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_neuprint}
\alias{manc_neuprint}
\title{Login to MANC neuprint server}
\usage{
manc_neuprint(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{neuprint_login}}}
}
\value{
a \code{\link{neuprint_connection}} object returned by \code{\link{neuprint_login}}
}
\description{
Login to MANC neuprint server
}
\examples{
\donttest{
vncc=manc_neuprint()
anchorids <- neuprintr::neuprint_ids("status:Anchor", conn=vncc)
# the last connection will be used by default
anchormeta <- neuprintr::neuprint_get_meta("status:Anchor")

plot(cumsum(sort(anchormeta$pre, decreasing = T)), ylab='Cumulative presynapses')
plot(cumsum(sort(anchormeta$post, decreasing = T)), ylab='Cumulative postsynapses')

}
}
