% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_neuprint_meta}
\alias{manc_neuprint_meta}
\title{Fetch neuprint metadata for MANC neurons}
\usage{
manc_neuprint_meta(ids = NULL, conn = manc_neuprint(), roiInfo = FALSE)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}

\item{roiInfo}{whether to include the \code{roiInfo} field detailing synapse
numbers in different locations. This is omitted by default as it is
returned as a character vector of unprocessed JSON.}
}
\value{
A data.frame with one row for each (unique) id and NAs for all
  columns except bodyid when neuprint holds no metadata.
}
\description{
Fetch neuprint metadata for MANC neurons
}
\examples{
\donttest{
manc_neuprint_meta("DNp01")
}
}
