% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_neuprint_meta}
\alias{manc_neuprint_meta}
\title{Fetch neuprint metadata for MANC neurons}
\usage{
manc_neuprint_meta(ids, conn = manc_neuprint())
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}
}
\value{
A data.frame with one row for each (unique) id and NAs for all
  columns except bodyid when neuprint holds no metadata.
}
\description{
Fetch neuprint metadata for MANC neurons
}
\examples{
\donttest{
manc_neuprint_meta("Giant Fiber")
}
}
