% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{manc_read_neurons}
\alias{manc_read_neurons}
\title{Read MANC skeletons via neuprint}
\usage{
manc_read_neurons(
  ids,
  connectors = FALSE,
  heal.threshold = Inf,
  conn = manc_neuprint(),
  ...
)
}
\arguments{
\item{ids}{A set of body ids in any form understandable to
\code{\link{manc_ids}}}

\item{connectors}{whether or not to add synapse data to the retrieved
skeletons in the format used by the \code{rcatmaid} package, for easy use
with \code{rcatmaid} or \code{catnat} functions. This can be done for
synapse-less skeletons using \code{neuprint_assign_connectors}}

\item{heal.threshold}{distance in raw units beyond which isolated fragments
will not be merged onto the main skeleton. The default of \code{1000}
implies 8000 nm for the hemibrain dataset. Use \code{Inf} to merge all
fragments.}

\item{conn}{Optional, a \code{\link{neuprint_connection}} object, which also
specifies the neuPrint server. Defaults to \code{\link{manc_neuprint}()} to
ensure that query is against the VNC dataset.}

\item{...}{Additional arguments passed to
\code{neuprintr::\link{neuprint_read_neurons}}.}
}
\value{
A \code{\link{neuronlist}} with attached metadata
}
\description{
Read MANC skeletons via neuprint
}
\details{
\code{manc_read_neurons} fetches metadata from neuprint but does not
  fetch synapse locations by default as this is very time consuming.
}
\examples{
\dontrun{
gfs=manc_read_neurons("Giant Fiber")
gfs[,]
}
}
