% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{manc_scene}
\alias{manc_scene}
\title{Return a Neuroglancer scene URL for MANC dataset}
\usage{
manc_scene(
  ids = NULL,
  node = "clio",
  open = FALSE,
  basescene = c("2022-04-13", "2022-01-13", "2021-05-05", "2021-05-04", "2021-04-01",
    "2021-02-01"),
  server = c("clio", "appspot", "janelia"),
  return.json = FALSE
)
}
\arguments{
\item{ids}{A set of body ids to add to the neuroglancer scene in any form
compatible with \code{\link{manc_ids}}}

\item{node}{A DVID node e.g. as returned by \code{manc_dvid_node}. The
(recommended) default behaviour is to use the current Clio node.}

\item{open}{When \code{TRUE} opens the URL in your browser.}

\item{basescene}{Which neuroglancer scene url to use as a base. You can also
supply your own URL.}

\item{server}{Whether to use Janelia's Clio branch, the Google server (newest
version of neuroglancer) or the Janelia server (required for annotation in
early 2021, but now deprecated in favour of Clio). 99% of the time you
should keep the default.}

\item{return.json}{Whether to return a JSON fragment defining the scene or
(by default) a Neuroglancer URL.}
}
\value{
A character vector containing a single Neuroglancer URL or a JSON
  fragment.
}
\description{
The default behaviour is to generate a rich neuroglancer scene
  with including any passed \code{ids} using the current Clio DVID node. This
  means that meshes should be in sync. See
  \href{https://flyem-cns.slack.com/archives/C01MYQ1AQ5D/p1619825198227100?thread_ts=1619816902.216600&cid=C01MYQ1AQ5D}{this
   slack} post from Stuart Berg for more details.
}
\details{
Neuroglancer scenes can be pasted into a variety of different
  variants. Use the \code{return.json} to get a JSON fragment that can be
  pasted into any neuroglancer instance using the closed curly bracket
  symbol.

  See
  \href{https://flyem-cns.slack.com/archives/C01MYQ1AQ5D/p1624033684227400}{slack}
  for why \url{https://clio-ng.janelia.org/} is the recommended base Url
  (chosen when \code{server='clio'}).
}
\section{scenes}{
 The following scenes (named by the approximate date that we
  started using them) are available. \itemize{

\item \code{2022-04-13} Fixed synapses display and segmentation layer v.0.68.
\item \code{2021-05-05} Like \code{2021-05-04} but with a voxelwise ROI
  segmentation layer copied over from \code{2021-04-01}.
\item \code{2021-05-04} Added nerves and full VNC (cell body rind) surface
mesh. See
\href{https://flyem-cns.slack.com/archives/C01BT2XFEEN/p1620164087077600}{Slack
message from Stuart Berg}. GSXEJ added the ROIs to

\item \code{2021-04-01} With VNC ROIs. By April 2021 we were using
\href{https://clio.janelia.org/}{Clio} for annotations.

\item \code{2021-02-01} In early 2021 we were using a Janelia server hosting
neuroglancer that allowed annotation through a hybrid DVID backend.

The early 2021 server required authentication at
\url{https://neuprint.janelia.org/} in order to use the annotation features.
I recommend logging in and out of neuprint if you still get authentication
errors from Neuroglancer when attempting to use the annotation layer.

}
}

\examples{
\dontrun{
browseURL(manc_scene())
# copy scene information with a sample neuron to the clipboard
clipr::write_clip(manc_scene(ids=13749))

# JSON fragment that could be copied into Clio
clipr::write_clip(manc_scene(return.json = TRUE))

}
}
