% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{manc_user_annotations}
\alias{manc_user_annotations}
\title{Read point annotations from DVID using neuprint authentication}
\usage{
manc_user_annotations(
  email = "jefferis@gmail.com",
  node = "36e0b",
  raw = FALSE,
  neuprint_connection = NULL,
  tz = "UTC"
)
}
\arguments{
\item{email}{The google email address used to authenticate with neuprint}

\item{node}{The DVID node identifier}

\item{raw}{Whether to return the raw \code{httr::\link{GET}} response
(default \code{FALSE}).}

\item{neuprint_connection}{A \code{\link{neuprintr}} connection object
returned by \code{neuprint_login}. This includes the required authorisation
information to connect to DVID.}

\item{tz}{Time zone for edit timestamps. Defaults to "UTC" i.e. Universal
Time, Coordinated. Set to "" for your current timezone. See
\code{\link{as.POSIXct}} for more details.}

\item{simplifyVector}{Whether to simplify lists to vectors (and data frames
where appropriate). Default \code{TRUE}, see
\code{jsonlite::\link{fromJSON}} for details.}
}
\value{
A data.frame
}
\description{
Read point annotations from DVID using neuprint authentication
}
\examples{
\donttest{
df=manc_user_annotations()
head(df)
json=httr::content(manc_read_annotations(raw=TRUE))
}
}
