% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzpos.R
\name{manc_xyz2bodyid}
\alias{manc_xyz2bodyid}
\title{Find the bodyid for an XYZ location}
\usage{
manc_xyz2bodyid(xyz, node = "neutu", cache = FALSE)
}
\arguments{
\item{xyz}{location in raw MANC pixels}

\item{node}{A DVID node as returned by \code{\link{manc_dvid_node}}. The
default is to return the current active (unlocked) node being used through
neutu.}

\item{cache}{Whether to cache the result of this call for 5 minutes.}
}
\value{
A character vector of body ids (0 is missing somas / missing locations)
}
\description{
Find the bodyid for an XYZ location
}
\examples{
ids=manc_xyz2bodyid(mancneckseeds)
tids=table(ids)
# are there many duplicate ids
table(tids)

dups=names(tids[tids>1])
\dontrun{
manc_scene(ids=dups)
}
}
