% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mancsomapos}
\alias{mancsomapos}
\title{Soma locations for all intrinsic neurons of the ventral nerve cord}
\format{
An object of class \code{data.frame} with 15613 rows and 8 columns.
}
\usage{
mancsomapos
}
\description{
These were generated as neuroglancer annotations by Chris Ordish of the FlyEM
team and
\href{https://flyem-cns.slack.com/archives/C01BT2XFEEN/p1614781577010700}{passed
to Greg Jefferis on 3rd March 2021 by Pat Rivlin}.
}
\details{
Note that these positions are raw (8nm voxel) locations.
}
\examples{
head(mancsomapos)

library(nat)
points3d(xyzmatrix(mancsomapos))
plot3d(boundingbox(xyzmatrix(mancsomapos)))

# nm
summary(xyzmatrix(mancsomapos)*8)
# microns
summary(xyzmatrix(mancsomapos)/125)
}
\keyword{datasets}
